+++ frontend_tests/zjsunit/index.js
@@ -39,6 +39,10 @@ global.write_handlebars_output = output.write_handlebars_output;
 global.write_test_output = output.write_test_output;
 global.append_test_output = output.append_test_output;
 
+// Set up fake jQuery
+var zjquery = require('./zjquery.js');
+global.zjquery = zjquery.zjquery;
+
 var noop = function () {};
 
 output.start_writing();
